# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Signal87 Core API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /documents/multi-chat:
    post:
      operationId: multiChat
      tags: [chat]
      summary: Ask one question across 2-5 selected documents
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MultiChatInput"
      responses:
        "200":
          description: Synthesized answer with grouped citations and verification trace
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MultiChatResult"
        "400":
          description: Invalid selection (fewer than 2 or more than 5 documents, or a selected document has no chunks)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: One or more selected documents not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /trash:
    get:
      operationId: listTrash
      tags: [trash]
      summary: List trashed documents with pagination
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Trashed documents
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrashList"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /trash/{id}/restore:
    post:
      operationId: restoreTrash
      tags: [trash]
      summary: Restore a trashed document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Restored
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrashRestore"
        "404":
          description: Document not found in trash
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /trash/{id}:
    delete:
      operationId: permanentDelete
      tags: [trash]
      summary: Permanently delete a trashed document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Permanently deleted
        "404":
          description: Document not found in trash
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /trash/empty:
    post:
      operationId: emptyTrash
      tags: [trash]
      summary: Empty all trash
      responses:
        "200":
          description: Trash emptied
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrashEmpty"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    MultiChatInput:
      type: object
      required: [documentIds, question]
      properties:
        documentIds:
          type: array
          minItems: 2
          maxItems: 5
          items:
            type: integer
        question:
          type: string
          minLength: 1
    MultiCitation:
      type: object
      required: [citationNumber, documentId, documentName, chunkIndex, content, relevanceScore]
      properties:
        citationNumber:
          type: integer
        documentId:
          type: integer
        documentName:
          type: string
        chunkIndex:
          type: integer
        content:
          type: string
        relevanceScore:
          type: number
    MultiChatDocumentStat:
      type: object
      required: [documentId, documentName, chunksSearched, chunksRetrieved]
      properties:
        documentId:
          type: integer
        documentName:
          type: string
        chunksSearched:
          type: integer
        chunksRetrieved:
          type: integer
    MultiDebugInfo:
      type: object
      required:
        - route
        - provider
        - model
        - fallbackUsed
        - documentIds
        - documentNames
        - documentsSearched
        - chunksSearched
        - chunksRetrieved
        - chunksRetrievedByDocument
        - retrievalLatencyMs
        - llmLatencyMs
        - totalLatencyMs
      properties:
        route:
          type: string
        provider:
          type: string
        model:
          type: string
        fallbackUsed:
          type: boolean
        documentIds:
          type: array
          items:
            type: integer
        documentNames:
          type: array
          items:
            type: string
        documentsSearched:
          type: integer
        chunksSearched:
          type: integer
        chunksRetrieved:
          type: integer
        chunksRetrievedByDocument:
          type: array
          items:
            $ref: "#/components/schemas/MultiChatDocumentStat"
        retrievalLatencyMs:
          type: number
        llmLatencyMs:
          type: number
        totalLatencyMs:
          type: number
        errors:
          type: ["string", "null"]
    MultiChatResult:
      type: object
      required: [answer, citations, debug]
      properties:
        answer:
          type: string
        citations:
          type: array
          items:
            $ref: "#/components/schemas/MultiCitation"
        debug:
          $ref: "#/components/schemas/MultiDebugInfo"
...
    TrashItem:
      type: object
      required: [id, fileName, fileType, fileSize, uploadedAt, deletedAt, extractionStatus, chunkCount, originalFileAvailable]
      properties:
        id:
          type: integer
        fileName:
          type: string
        fileType:
          type: string
        fileSize:
          type: ["integer", "null"]
        uploadedAt:
          type: string
          format: date-time
        deletedAt:
          type: ["string", "null"]
          format: date-time
        extractionStatus:
          type: string
        chunkCount:
          type: integer
        originalFileAvailable:
          type: boolean
    TrashList:
      type: object
      required: [items, total, limit, offset]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/TrashItem"
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
    TrashRestore:
      type: object
      required: [id, fileName, fileType, restored]
      properties:
        id:
          type: integer
        fileName:
          type: string
        fileType:
          type: string
        restored:
          type: boolean
    TrashEmpty:
      type: object
      required: [deletedCount]
      properties:
        deletedCount:
          type: integer
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string