# lib/api-zod/src/generated/types/trashItem.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Signal87 Core API specification
 * OpenAPI spec version: 0.1.0
 */

export interface TrashItem {
  id: number;
  fileName: string;
  fileType: string;
  /** @nullable */
  fileSize: number | null;
  uploadedAt: Date;
  /** @nullable */
  deletedAt: Date | null;
  extractionStatus: string;
  chunkCount: number;
  originalFileAvailable: boolean;
}