# lib/api-zod/src/generated/types/trashList.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Signal87 Core API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TrashItem } from './trashItem';

export interface TrashList {
  items: TrashItem[];
  total: number;
  limit: number;
  offset: number;
}