# lib/api-zod/src/generated/types/trashRestore.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Signal87 Core API specification
 * OpenAPI spec version: 0.1.0
 */

export interface TrashRestore {
  id: number;
  fileName: string;
  fileType: string;
  restored: boolean;
}